@@ -20,27 +20,34 @@
         "issues": "https://github.com/sebastianbergmann/php-code-coverage/issues"
     },
     "config": {
+        "platform": {
+            "php": "7.3.0"
+        },
         "optimize-autoloader": true,
         "sort-packages": true
     },
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2",
+        "php": ">=7.3",
         "ext-dom": "*",
+        "ext-libxml": "*",
         "ext-xmlwriter": "*",
-        "phpunit/php-file-iterator": "^2.0.2",
-        "phpunit/php-token-stream": "^3.1.3 || ^4.0",
-        "phpunit/php-text-template": "^1.2.1",
-        "sebastian/code-unit-reverse-lookup": "^1.0.1",
-        "sebastian/environment": "^4.2.2",
-        "sebastian/version": "^2.0.1",
-        "theseer/tokenizer": "^1.1.3"
+        "nikic/php-parser": "^4.13.0",
+        "phpunit/php-file-iterator": "^3.0.3",
+        "phpunit/php-text-template": "^2.0.2",
+        "sebastian/code-unit-reverse-lookup": "^2.0.2",
+        "sebastian/complexity": "^2.0",
+        "sebastian/environment": "^5.1.2",
+        "sebastian/lines-of-code": "^1.0.3",
+        "sebastian/version": "^3.0.1",
+        "theseer/tokenizer": "^1.2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.2.2"
+        "phpunit/phpunit": "^9.3"
     },
     "suggest": {
-        "ext-xdebug": "^2.7.2"
+        "ext-pcov": "*",
+        "ext-xdebug": "*"
     },
     "autoload": {
         "classmap": [
@@ -55,7 +62,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "7.0-dev"
+            "dev-master": "9.2-dev"
         }
     }
 }
