@@ -18,11 +18,19 @@
     "support": {
         "issues": "https://github.com/sebastianbergmann/php-file-iterator/issues"
     },
+    "config": {
+        "platform": {
+            "php": "7.3.0"
+        },
+        "optimize-autoloader": true,
+        "sort-packages": true
+    },
+    "prefer-stable": true,
     "require": {
-        "php": ">=7.1"
+        "php": ">=7.3"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^9.3"
     },
     "autoload": {
         "classmap": [
@@ -31,7 +39,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.0.x-dev"
+            "dev-master": "3.0-dev"
         }
     }
 }
