@@ -21,7 +21,7 @@
     },
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2",
+        "php": ">=7.3",
         "ext-dom": "*",
         "ext-json": "*",
         "ext-libxml": "*",
@@ -29,36 +29,39 @@
         "ext-xml": "*",
         "ext-xmlwriter": "*",
         "doctrine/instantiator": "^1.3.1",
-        "myclabs/deep-copy": "^1.10.0",
+        "myclabs/deep-copy": "^1.10.1",
         "phar-io/manifest": "^2.0.3",
         "phar-io/version": "^3.0.2",
-        "phpspec/prophecy": "^1.10.3",
-        "phpunit/php-code-coverage": "^7.0.12",
-        "phpunit/php-file-iterator": "^2.0.4",
-        "phpunit/php-text-template": "^1.2.1",
-        "phpunit/php-timer": "^2.1.2",
-        "sebastian/comparator": "^3.0.2",
-        "sebastian/diff": "^3.0.2",
-        "sebastian/environment": "^4.2.3",
-        "sebastian/exporter": "^3.1.2",
-        "sebastian/global-state": "^3.0.0",
-        "sebastian/object-enumerator": "^3.0.3",
-        "sebastian/resource-operations": "^2.0.1",
-        "sebastian/type": "^1.1.3",
-        "sebastian/version": "^2.0.1"
+        "phpspec/prophecy": "^1.12.1",
+        "phpunit/php-code-coverage": "^9.2.13",
+        "phpunit/php-file-iterator": "^3.0.5",
+        "phpunit/php-invoker": "^3.1.1",
+        "phpunit/php-text-template": "^2.0.3",
+        "phpunit/php-timer": "^5.0.2",
+        "sebastian/cli-parser": "^1.0.1",
+        "sebastian/code-unit": "^1.0.6",
+        "sebastian/comparator": "^4.0.5",
+        "sebastian/diff": "^4.0.3",
+        "sebastian/environment": "^5.1.3",
+        "sebastian/exporter": "^4.0.3",
+        "sebastian/global-state": "^5.0.1",
+        "sebastian/object-enumerator": "^4.0.3",
+        "sebastian/resource-operations": "^3.0.3",
+        "sebastian/type": "^2.3.4",
+        "sebastian/version": "^3.0.2"
     },
     "require-dev": {
-        "ext-PDO": "*"
+        "ext-PDO": "*",
+        "phpspec/prophecy-phpunit": "^2.0.1"
     },
     "config": {
         "platform": {
-            "php": "7.2.0"
+            "php": "7.3.0"
         },
         "optimize-autoloader": true,
         "sort-packages": true
     },
     "suggest": {
-        "phpunit/php-invoker": "^2.0.0",
         "ext-soap": "*",
         "ext-xdebug": "*"
     },
@@ -68,6 +71,9 @@
     "autoload": {
         "classmap": [
             "src/"
+        ],
+        "files": [
+            "src/Framework/Assert/Functions.php"
         ]
     },
     "autoload-dev": {
@@ -75,7 +81,6 @@
             "tests/"
         ],
         "files": [
-            "src/Framework/Assert/Functions.php",
             "tests/_files/CoverageNamespacedFunctionTest.php",
             "tests/_files/CoveredFunction.php",
             "tests/_files/NamespaceCoveredFunction.php"
@@ -83,7 +88,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "8.5-dev"
+            "dev-master": "9.5-dev"
         }
     }
 }
