@@ -2,7 +2,7 @@
     "name": "sebastian/exporter",
     "description": "Provides the functionality to export PHP variables for visualization",
     "keywords": ["exporter","export"],
-    "homepage": "http://www.github.com/sebastianbergmann/exporter",
+    "homepage": "https://www.github.com/sebastianbergmann/exporter",
     "license": "BSD-3-Clause",
     "authors": [
         {
@@ -27,16 +27,19 @@
         }
     ],
     "config": {
+        "platform": {
+            "php": "7.3.0"
+        },
         "optimize-autoloader": true,
         "sort-packages": true
     },
     "prefer-stable": true,
     "require": {
-        "php": ">=7.0",
-        "sebastian/recursion-context": "^3.0"
+        "php": ">=7.3",
+        "sebastian/recursion-context": "^4.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5",
+        "phpunit/phpunit": "^9.3",
         "ext-mbstring": "*"
     },
     "autoload": {
@@ -46,7 +49,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "3.1.x-dev"
+            "dev-master": "4.0-dev"
         }
     }
 }
