@@ -1,68 +1,83 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace App\Core;
 
+use App\Container\Exceptions\ContainerException;
+use App\Helpers\EnglishDictionary;
+use App\Interfaces\DictionaryInterface;
+use App\Interfaces\PlayerInterface;
+use App\Models\Player;
+use App\Container\Container;
 use Exception;
 
 class Router
 {
-    public $routes = [
+    public array $routes = [
         'GET' => [],
         'POST' => [],
     ];
 
-    /**
-     * @param $uri
-     * @param $controller
-     * @return mixed
-     */
-    public function get($uri, $controller, $options = [])
+    private Container $container;
+
+    public function __construct(Container $container)
     {
-        return $this->routes['GET'][$uri] = [
+        $this->container = $container;
+        $container->set(DictionaryInterface::class, EnglishDictionary::class);
+        $container->set(PlayerInterface::class, Player::class);
+    }
+
+    /**
+     * @param string $uri
+     * @param string $controller
+     * @param array $options
+     */
+    public function get(string $uri, string $controller, array $options = [])
+    {
+        $this->routes['GET'][$uri] = [
             'controller' => $controller,
-            'options' => $options
+            'options' => $options,
         ];
     }
 
     /**
      * @param $uri
      * @param $controller
-     * @return mixed
      */
     public function post($uri, $controller, $options = [])
     {
-        return $this->routes['POST'][$uri] = [
+        $this->routes['POST'][$uri] = [
             'controller' => $controller,
-            'options' => $options
+            'options' => $options,
         ];
     }
 
     /**
      * @throws Exception
      */
-    public function direct($uri, $type)
+    public function direct(string $uri, string $type)
     {
         if (!array_key_exists($uri, $this->routes[$type])) {
-            throw new Exception('No Route');
+            // todo: create 404 page
+            http_response_code(404);
+            die;
         }
-            // todo: inject dependencies
-        $this->checkAuthUser($type, $uri);
 
+        $this->checkAuthUser($type, $uri);
         $data = explode('@', $this->routes[$type][$uri]['controller']);
 
         return $this->callAction($data[0], $data[1]);
     }
 
     /**
-     * @param $controller
-     * @param $method
+     * @param string $controller
+     * @param string $method
      * @return mixed
-     * @throws Exception
+     * @throws ContainerException
      */
-    private function callAction($controller, $method)
+    private function callAction(string $controller, string $method): mixed
     {
         $controller = 'App\Controllers\\' . $controller;
-        $controller = new $controller;
+        $controller = $this->container->get($controller);
 
         if (!method_exists($controller, $method)) {
             throw new Exception("{$method} does not belong to {$controller}");
