@@ -1,63 +1,74 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace App\Models;
 
+use App\Core\App;
 use Exception;
 
 class Word
 {
-    private $word = '';
     private $config;
     private $database;
 
-    public function __construct(array $config, $database = null, string $word)
+    /**
+     * @throws Exception
+     */
+    public function __construct(App $app)
     {
-        $this->config = $config;
-        $this->word = $word;
-        $this->database = $database;
+        $this->config = $app::get('config');
+        $this->database = $app::get('database');
     }
     /**
      * @return void
      */
-    public function insert()
+    public function insert($word)
     {
         try {
-            $points = $this->calculatePoints();
+            $points = $this->calculatePoints($word);
 
             return $this->database->insert('words', [
-                'word' => $this->word,
+                'word' => $word,
                 'user_id' => $_SESSION['user']['id'],
                 'points' => $points['total'],
                 'is_palindrome' => $points['status']
             ]);
         } catch (Exception $e) {
             echo $e->getMessage();
+            die;
         }
     }
 
-    public function getWord()
+    public function getWord($word)
     {
-        // todo: prepare statement
-        return $this->database->get('SELECT id FROM words WHERE word = "' . $this->word . '" AND user_id = ' . $_SESSION['user']['id']);
+        $sql = [
+            'sql' => 'SELECT id FROM words WHERE word = :word AND user_id = :user_id',
+            'params' => [
+                'word' => $word,
+                'user_id' => $_SESSION['user']['id']
+                ]
+            ];
+
+        return $this->database->get($sql);
     }
 
     /**
      * @param $word
      * @return bool
      */
-    private function isPalindrome()
+    private function isPalindrome($word): bool
     {
-        return $this->word == strrev($this->word);
+        return $word == strrev($word);
     }
 
     /**
+     * @param $word
      * @return bool
      */
-    private function isAlmostPalindrome()
+    private function isAlmostPalindrome($word): bool
     {
-        $reverse = strrev($this->word);
-        $length = strlen($this->word);
-        $sim = similar_text($this->word, $reverse);
+        $reverse = strrev($word);
+        $length = strlen($word);
+        $sim = similar_text($word, $reverse);
 
         if ($length - 1 == $sim) {
             return true;
@@ -67,17 +78,17 @@ class Word
     }
 
     /**
+     * @param $word
      * @return array
-     * @throws Exception
      */
-    private function calculatePalindromePoints()
+    private function calculatePalindromePoints($word): array
     {
-        if ($this->isPalindrome()) {
+        if ($this->isPalindrome($word)) {
             return [
                 'points' => (int)$this->config['palindrome_points'],
                 'status' => 1
             ];
-        } else if ($this->isAlmostPalindrome()) {
+        } else if ($this->isAlmostPalindrome($word)) {
             return [
                 'points' => (int)$this->config['almost_palindrome_points'],
                 'status' => 2
@@ -95,10 +106,10 @@ class Word
      * @return array
      * @throws Exception
      */
-    public function calculatePoints()
+    public function calculatePoints($word): array
     {
-        $palPoints = $this->calculatePalindromePoints();
-        $wordSplit = str_split($this->word);
+        $palPoints = $this->calculatePalindromePoints($word);
+        $wordSplit = str_split($word);
         $total = count(array_unique($wordSplit));
         $total += $palPoints['points'];
 
