@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace App\Core;
 
@@ -6,17 +6,25 @@ use Exception;
 
 class App
 {
-    protected static $container = [];
+    protected static array $container = [];
 
+    /**
+     * @return void
+     */
     public static function bind($key, $value)
     {
         static::$container[$key] = $value;
     }
 
-    public static function get($key)
+    /**
+     * @param string $key
+     * @return mixed
+     * @throws Exception
+     */
+    public static function get(string $key): mixed
     {
         if (!array_key_exists($key, static::$container)) {
-            throw new Exception("No {$key} in the container");
+            throw new Exception("No {$key} in the App container");
         }
 
         return static::$container[$key];
