@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace App\Core;
 
@@ -6,15 +6,15 @@ class Request {
     /**
      * @return string
      */
-    public static function uri()
+    public static function uri(): string
     {
         return trim(parse_url($_SERVER['REQUEST_URI'],PHP_URL_PATH),'/');
     }
 
     /**
-     * @return mixed
+     * @return string
      */
-    public static function method()
+    public static function method(): string
     {
         return $_SERVER['REQUEST_METHOD'];
     }
