@@ -1,14 +1,17 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace App\Core\Config;
+
+use Exception;
 use PDO;
+use PDOException;
 
 class Connection
 {
     /**
      * @throws Exception
      */
-    public static function connect($config)
+    public static function connect($config): PDO
     {
         try {
             return new PDO(
