@@ -1,33 +1,96 @@
 <?php declare(strict_types=1);
 
-use App\app\helpers\UserValidation;
-use App\Models\User;
+use App\Helpers\UserValidation;
+use App\Controllers\UsersController;
+use App\Models\Player;
 use PHPUnit\Framework\TestCase;
 
 final class UserTest extends TestCase
 {
     /**
      * @test
+     * @dataProvider users
      */
-    public function checkUserValidation()
+    public function checkUserValidation($status, $data)
     {
-        $mockRepo = $this->createMock(User::class);
-        $getEmail = $mockRepo->getEmail('test@test.com');
+        $mockPlayer = $this->createMock(Player::class);
+        $_POST = $data;
+        $validation = new UserValidation($mockPlayer);
 
-        $mockRepo->method('getEmail')->willReturn('test@test.com');
+        $this->assertEquals($status, $validation->validate());
+    }
 
+    /**
+     * @test
+     */
+    public function createUser()
+    {
+        $mockPlayer = $this->createMock(Player::class);
+        $mockPlayer->method('create')->willReturn(true);
+        $mockValidator = $this->createMock(UserValidation::class);
+        $mockValidator->method('validate')->willReturn(false);
 
-        $user = [
-            'first_name' => 'Mario',
-            'last_name' => 'Stanic',
-            'email' => 'mario90stanic@gmail.com',
+        $userController = new UsersController($mockValidator, $mockPlayer);
+
+        $_POST = [
+            'first_name' => 'test_first_name',
+            'last_name' => 'test_last_name',
+            'email' => 'test@test.com',
             'password' => '123456',
             'repeat_password' => '123456',
         ];
-//
-        $validation = new UserValidation($user);
-        $this->assertTrue(true);
-//
-        $this->assertTrue($validation->validate());
+
+        $this->assertTrue($userController->store());
+
+    }
+
+    public function users()
+    {
+        return [
+            [
+                'status' => true,
+                'data' => [
+                    'id' => 1,
+                    'first_name' => 'test_name',
+                    'last_name' => '',
+                    'email' => 'test@test.com',
+                    'password' => '123456',
+                    'repeat_password' => '123456',
+                ]
+            ],
+            [
+                'status' => true,
+                'data' => [
+                    'id' => 1,
+                    'first_name' => 'test_first_name',
+                    'last_name' => 'test_last_name',
+                    'email' => '',
+                    'password' => '123456',
+                    'repeat_password' => '123456',
+                ]
+            ],
+            [
+                'status' => true,
+                'data' => [
+                    'id' => 1,
+                    'first_name' => 'test_first_name',
+                    'last_name' => 'test_last_name',
+                    'email' => 'test@test.com',
+                    'password' => '',
+                    'repeat_password' => '123456',
+                ]
+            ],
+            [
+                'status' => false,
+                'data' => [
+                    'id' => 1,
+                    'first_name' => 'test_first_name',
+                    'last_name' => 'test_last_name',
+                    'email' => 'test@test.com',
+                    'password' => '123456',
+                    'repeat_password' => '123456',
+                ]
+            ]
+        ];
     }
 }
